Move LogoutButton to a component. Add alert when login fails. #10
