+++ src/components/Page.tsx
@@ -1,13 +1,13 @@
 import '../App.css'
 import { Box, Tabs, AppBar } from '@mui/material'
 import React from 'react'
-import LogoutIcon from '@mui/icons-material/Logout'
 import Button from '@mui/material/Button'
 import Menu from '@mui/material/Menu'
 import MenuItem from '@mui/material/MenuItem'
 import MenuIcon from '@mui/icons-material/Menu'
 import { useMobileScreen } from '../hooks/useMobileScreen'
-import { Link, useNavigate } from 'react-router-dom'
+import { Link } from 'react-router-dom'
+import LogoutButton from './LogoutButton'
 
 interface PageProps {
     children?: React.ReactNode
@@ -30,12 +30,6 @@ function Page(props: PageProps) {
     }
 
     const path = window.location.pathname
-    const navigate = useNavigate()
-
-    const handleLogout = () => {
-        window.localStorage.removeItem('AuthPayload')
-        navigate('/')
-    }
 
     return (
         <Box sx={{ width: '100%' }}>
@@ -136,28 +130,7 @@ function Page(props: PageProps) {
                     </Tabs>
                 )}
 
-                <button
-                    onClick={handleLogout}
-                    style={{
-                        display: 'flex',
-                        flexDirection: 'row',
-                        alignItems: 'center',
-                        backgroundColor: 'transparent',
-                        border: 'none',
-                        cursor: 'pointer',
-                    }}
-                >
-                    <p style={{ fontSize: 15, color: 'white' }}>
-                        Kirjaudu ulos
-                    </p>
-                    <LogoutIcon
-                        style={{
-                            marginRight: 15,
-                            marginLeft: 10,
-                            color: 'white',
-                        }}
-                    />
-                </button>
+                <LogoutButton />
             </AppBar>
             {props.children}
         </Box>
